+++ utils.py
@@ -34,3 +34,22 @@ def ignored(*exceptions):
         yield
     except exceptions:
         pass
+
+
+class Observable(object):
+    def __init__(self):
+        self._observers = []
+
+    def add_observer(self, observer):
+        if not observer in self._observers:
+            self._observers.append(observer)
+
+    def remove_observer(self, observer):
+        try:
+            self._observers.remove(observer)
+        except ValueError:
+            pass
+
+    def notify_observers(self):
+        for observer in self._observers:
+            observer.update()
